pyproject.toml
[tool.poetry]
name = "satori"
version = "0.1.0"
description = ""
authors = ["Roberto Garcia <robertojesusgarcia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"
pre-commit = "^3.6.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
retrying = "^1.3.4"
secure = "^0.3.0"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
